Divided responsibily of MazeGenerator
