+++ public/mazeTests.html
@@ -7,7 +7,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Maze Runners</title>
-    <link rel="stylesheet" href="./src/styles/maze-tests-style.css">
+    <link rel="stylesheet" href="./src/styles/mazeTests.css">
 </head>
 <body>
     <div id="maze">
