+++ public/src/scripts/classes/PathCreator.js
@@ -1,0 +1,4 @@
+export class PathCreator
+{
+    
+}
