+++ public/src/scripts/runMaze.js
@@ -1,4 +1,11 @@
-import { MazeGenerator } from "./classes/mazeGenerator.js";
+import { MatrixGenerator } from "./classes/MatrixGenerator.js";
+import { MazeRenderizer } from "./classes/MazeRenderizer.js";
+import { PathCreator } from "./classes/PathCreator.js";
 
-const maze = new MazeGenerator(13, 13, 2);
-maze.generateMaze();
+const mazeMatrix = new MatrixGenerator(13, 13, 2);
+const matrix = mazeMatrix.getMatrix();
+
+
+
+const maze = new MazeRenderizer(matrix);
+maze.renderizeMaze();
