+++ public/src/styles/mazeTests.css
@@ -14,7 +14,7 @@
 {
     display: flex;
     flex-direction: column;
-    border: solid 1px;
+    border: solid 2px;
     width: 30px;
     height: 30px;
 }
